requirements.txt
fire
databricks-sql-connector
requests